# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CPCBusiness Enterprise Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: getTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
        - name: assigneeId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── TEAM ────────────────────────────────────────────────────────────────────
  /team:
    get:
      operationId: getTeamMembers
      tags: [team]
      summary: List team members
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamMember"
    post:
      operationId: createTeamMember
      tags: [team]
      summary: Create team member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamMemberInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
  /team/{id}:
    patch:
      operationId: updateTeamMember
      tags: [team]
      summary: Update team member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamMemberUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
    delete:
      operationId: deleteTeamMember
      tags: [team]
      summary: Delete team member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── INVOICES ────────────────────────────────────────────────────────────────
...
  /tickets:
    get:
      operationId: getTickets
      tags: [tickets]
      summary: List tickets
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: clientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"
    post:
      operationId: createTicket
      tags: [tickets]
      summary: Create ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{id}:
    get:
      operationId: getTicket
      tags: [tickets]
      summary: Get ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
    patch:
      operationId: updateTicket
      tags: [tickets]
      summary: Update ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"

  # ─── SERVICES CMS ────────────────────────────────────────────────────────────
  /services:
    get:
      operationId: getServices
      tags: [services]
      summary: List services
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      summary: Create service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    patch:
      operationId: updateService
      tags: [services]
      summary: Update service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── PORTFOLIO ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ─── AUTH SCHEMAS ───────────────────────────────────────────────────────
...
    Task:
      type: object
      required: [id, title, status, priority, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        projectId:
          type: ["integer", "null"]
        projectName:
          type: ["string", "null"]
        assigneeId:
          type: ["integer", "null"]
        assigneeName:
          type: ["string", "null"]
        status:
          type: string
        priority:
          type: string
        dueDate:
          type: ["string", "null"]
        createdAt:
          type: string
    TaskInput:
      type: object
      required: [title, status, priority]
      properties:
        title:
          type: string
        description:
          type: string
        projectId:
          type: integer
        assigneeId:
          type: integer
        status:
          type: string
        priority:
          type: string
        dueDate:
          type: string
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: string
        assigneeId:
          type: integer
        dueDate:
          type: string

    # ─── TEAM SCHEMAS ───────────────────────────────────────────────────────
    TeamMember:
      type: object
      required: [id, name, email, role, department, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        department:
          type: string
        avatarUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        phone:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        activeProjects:
          type: ["integer", "null"]
        createdAt:
          type: string
    TeamMemberInput:
      type: object
      required: [name, email, role, department]
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
        department:
          type: string
        phone:
          type: string
        skills:
          type: array
          items:
            type: string
    TeamMemberUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
        department:
          type: string
        isActive:
          type: boolean
        phone:
          type: string
        skills:
          type: array
          items:
            type: string

    # ─── INVOICE SCHEMAS ────────────────────────────────────────────────────
...
    Ticket:
      type: object
      required: [id, subject, status, priority, createdAt]
      properties:
        id:
          type: integer
        subject:
          type: string
        description:
          type: string
        clientId:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]
        assigneeId:
          type: ["integer", "null"]
        assigneeName:
          type: ["string", "null"]
        status:
          type: string
        priority:
          type: string
        category:
          type: ["string", "null"]
        resolution:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    TicketInput:
      type: object
      required: [subject, description, priority]
      properties:
        subject:
          type: string
        description:
          type: string
        clientId:
          type: integer
        priority:
          type: string
        category:
          type: string
    TicketUpdate:
      type: object
      properties:
        status:
          type: string
        priority:
          type: string
        assigneeId:
          type: integer
        resolution:
          type: string

    # ─── SERVICE CMS SCHEMAS ────────────────────────────────────────────────
    Service:
      type: object
      required: [id, title, category, isActive, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        category:
          type: string
        description:
          type: string
        shortDescription:
          type: ["string", "null"]
        iconName:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        features:
          type: array
          items:
            type: string
        technologies:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
    ServiceInput:
      type: object
      required: [title, category, description]
      properties:
        title:
          type: string
        category:
          type: string
        description:
          type: string
        shortDescription:
          type: string
        iconName:
          type: string
        imageUrl:
          type: string
        features:
          type: array
          items:
            type: string
        technologies:
          type: array
          items:
            type: string
        sortOrder:
          type: integer
    ServiceUpdate:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        description:
          type: string
        shortDescription:
          type: string
        iconName:
          type: string
        imageUrl:
          type: string
        features:
          type: array
          items:
            type: string
        technologies:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        sortOrder:
          type: integer

    # ─── PORTFOLIO SCHEMAS ──────────────────────────────────────────────────
...