# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CPCBusiness Enterprise Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  title: string;
  slug?: string;
  category: string;
  description?: string;
  /** @nullable */
  shortDescription?: string | null;
  /** @nullable */
  iconName?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  features?: string[];
  technologies?: string[];
  isActive: boolean;
  sortOrder?: number;
  createdAt: string;
}